Misterious patch to fix "Bug1" crash
With reclaiming-4.0 (GRUB rebased on 4.0), creating 7 CPU hogs and
scheduling them with SCHED_DEADLINE + reclaiming crashes the system in
a short time. The reason for this crash is still unknown...

This patch partly reverts to "reclaiming" (the 3.19-based version
of the thing) to fix the issue.
Some better investigation is still needed!