Add support for parallel reclaiming
Up to now, we tracked "per-runqueue utilisations", allowing to
perform the so-called "sequential reclaiming" (every runqueue can
reclaim its own inactive utilisation).
This commit adds support for tracking "per-root-domain" utilisation
and adds support for the so-called "parallel reclaiming" (tasks on
a runqueue can reclaim "global" inactive utilisation, taking advantage
of the runtime that has not been used on different runqueues).

The kind of reclaiming that we want can be configured through the
CONFIG_PARALLEL_RECLAIMING macro.

Signed-off-by: Luca Abeni <luca.abeni@unitn.it>